vidb/requirements.txt
azure-common==1.1.23
azure-keyvault==1.1.0
azure-nspkg==3.0.2
msrest==0.6.8
msrestazure==0.6.1

